optimize getUserWorkoutExercise

+++ FitnessServices/Models/UserWorkoutActivity.cs
@@ -8,7 +8,7 @@ public sealed class UserWorkoutActivityModel
     public long WorkoutExerciseId { get; set; }
     public int Set { get; set; }
     public int? Reps { get; set; }
-    public int? Time { get; set; }
+    public float? Time { get; set; }
     public float? Weight { get; set; }
     public bool? Saved { get; set; }
 }
